url2GSI: Don't redownload zip file

+++ url2GSI.sh
@@ -5,11 +5,14 @@ srctype=$2
 
 mkdir downloads
 
-wget -U "Mozilla/5.0" $url -O downloads/temp.zip
+if [ ! -f downloads/temp.zip ]; then
+    wget -U "Mozilla/5.0" $url -O downloads/temp.zip
+fi
+
 ./zip2img.sh downloads/temp.zip
 mkdir system
 mount cache/system.img system
 ./make.sh system $srctype AB
-#./make.sh system $srctype Aonly
+./make.sh system $srctype Aonly
 
 ls -ln out/
